requirements/dev.txt
-r common.txt
django-debug-toolbar==0.9.4
django_pdb==0.3.2
